include os

+++ test/hotspot/gtest/logging/test_logMessageTest.cpp
@@ -150,7 +150,7 @@ TEST_VM_F(LogMessageTest, long_message) {
   for (size_t i = strlen(start_marker); i < size; i++) {
     data[i] = '0' + (i % 10);
   }
-  int remaining_size = strlen(end_marker) + 1;
+  size_t remaining_size = strlen(end_marker) + 1;
   os::snprintf_checked(data + size - remaining_size, remaining_size, "%s", end_marker);
 
   msg.trace("%s", data); // Adds a newline, making the message exactly 10K in length.
